requirements.txt
langchain
langchain-core
langchain-community

groq
langchain-groq
langchain-openai
langchain-ollama

langchain-huggingface
langgraph
transformers

"unstructured[all-docs]" 
unstructured-client 
watermark

pytesseract 
pdf2image 
python-magic-bin
hf_xet

python-dotenv
pandas
numpy
fastapi
uvicorn
'pydantic[email]'
scikit-learn
presidio_analyzer 
presidio_anonymizer
deepdiff